+++ tests/snapshot_tests/test_snapshots.py
@@ -4781,3 +4781,22 @@ def test_mount_compose(snap_compare) -> None:
             await self.screen.mount_compose(compose_things())
 
     assert snap_compare(ComposeApp())
+
+
+def test_text_area_paste(snap_compare) -> None:
+    """Regression test for https://github.com/Textualize/textual/issues/4852
+
+    You should see a TextArea where the scrollbar and cursor at at the bottom
+    of the document.
+    """
+
+    class TextAreaApp(App):
+        def compose(self) -> ComposeResult:
+            yield TextArea()
+
+    async def run_before(pilot: Pilot) -> None:
+        await pilot.pause()
+        pilot.app._clipboard = "\n".join(["Where there is a Will"] * 100)
+        await pilot.press("ctrl+v")
+
+    assert snap_compare(TextAreaApp(), run_before=run_before)